requirements.txt
Flask
gunicorn
GoogleNews
datetime
newspaper3k

